Bringing back reverted overflow changes, one at a time

+++ clip.cpp
@@ -177,7 +177,7 @@ int clip(long long *x0, long long *y0, long long *x1, long long *y1, long long x
 		} else {
 			// failed both tests, so calculate the line segment to clip
 			// from an outside point to an intersection with clip edge
-			long long x = *x0, y = *y0;
+			__int128 x = *x0, y = *y0;
 
 			// At least one endpoint is outside the clip rectangle; pick it.
 			int outcodeOut = outcode0 ? outcode0 : outcode1;
@@ -187,16 +187,16 @@ int clip(long long *x0, long long *y0, long long *x1, long long *y1, long long x
 			// Now find the intersection point;
 			// use formulas y = y0 + slope * (x - x0), x = x0 + (1 / slope) * (y - y0)
 			if (outcodeOut & TOP) {	 // point is above the clip rectangle
-				x = *x0 + (*x1 - *x0) * (ymax - *y0) / (*y1 - *y0);
+				x = *x0 + (__int128) (*x1 - *x0) * (ymax - *y0) / (*y1 - *y0);
 				y = ymax;
 			} else if (outcodeOut & BOTTOM) {  // point is below the clip rectangle
-				x = *x0 + (*x1 - *x0) * (ymin - *y0) / (*y1 - *y0);
+				x = *x0 + (__int128) (*x1 - *x0) * (ymin - *y0) / (*y1 - *y0);
 				y = ymin;
 			} else if (outcodeOut & RIGHT) {  // point is to the right of clip rectangle
-				y = *y0 + (*y1 - *y0) * (xmax - *x0) / (*x1 - *x0);
+				y = *y0 + (__int128) (*y1 - *y0) * (xmax - *x0) / (*x1 - *x0);
 				x = xmax;
 			} else if (outcodeOut & LEFT) {	 // point is to the left of clip rectangle
-				y = *y0 + (*y1 - *y0) * (xmin - *x0) / (*x1 - *x0);
+				y = *y0 + (__int128) (*y1 - *y0) * (xmin - *x0) / (*x1 - *x0);
 				x = xmin;
 			}
